+++ config/tailwind/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @monext/tailwind-config
 
+## 1.1.1
+
+### Patch Changes
+
+- correct styles to tailwind css
+
 ## 1.1.0
 
 ### Minor Changes