+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @monext/ui
 
+## 0.0.3
+
+### Patch Changes
+
+- correct styles to tailwind css
+- Updated dependencies
+  - @monext/utils@0.0.1
+
 ## 0.0.2
 
 ### Patch Changes